INI parsing is now case insensitive
By default the code was case sensitive and was looking for options that
were starting with upercase letters. A prime example of where this
breaks is with magical camp. I do believe VXA defaults to exporting it
as all lower case so it is silly anyways to force this to be cahnged for
parsing.

* Make options and section names lower case when parsing the game ini
  file